feat(explorer): Add minimum load time for ripple effect
This change introduces a minimum load time of 1 second for the ripple loading animation in the Explorer view when in app mode and viewing a specific node. This ensures a smoother user experience by preventing the loader from disappearing too quickly. The animation duration for the ripple loader in the special application view has also been reduced for faster feedback.